+++ HISTORY.md
@@ -1,5 +1,14 @@
 # `Phys_782` Revision History
 
+## Revision 0.0.2
+
+- Setup basis and potentials directory
+- Added potential for bump down, sho and kp
+- Added unit tests
+- Run unit tests
+- setup Travis-CI
+- Add coverall
+
 ## Revision 0.0.1
 
 - Setup skelton repo directory